Module modus::options

Option valuation and betting optimization

Black-Scholes formula

Calculates the value of a European-type option using the Black-Scholes formula. Note that this is also valid for American-type call options but not for American-type put options, as shown by Merton (1973) provided the stock does not pay dividends. Because it uses the Black-Scholes formula, it has the same limitations, chiefly among them, the constant volatility

Usage:

 let a_option = Options{
    form: OptionType::Call,
    underlying: 43.0,
    strike: 55.0,
    maturity: 3,
    volatility: 0.7,
    rfr: 0.3,
    market_price: None,
 };
 println!("{}", bs_price(&a_option));

Monte-Carlo analysis

Alternatively, it performs a Monte-Carlo analysis to calculate the option price.

Usage:

let a_option = Options{
    form: OptionType::Call,
    underlying: 43.0,
    strike: 55.0,
    maturity: 3,
    volatility: 0.7,
    rfr: 0.3,
    market_price: None,
 };
 if let Ok(s) = expected(&a_option) { println!("{:?}", s); }

Kelly Criterion

If one were to be able to consistently find theoretical market values of the options different from their market values one could design an optimal strategy where the expected geometric growth rate is maximized by finding the fraction of the bankroll that maximizes the expected value of the logarithm of wealth, also known as the Kelly Criterion. This tries to implement the Kelly Criterion to find that fraction for a given option. I’m not fully certain the implementation is correct however, so you might want to consider a more mature crate for this.

Usage:

let a_option = Options{
    form: OptionType::Call,
    underlying: 43.0,
    strike: 55.0,
    maturity: 3,
    volatility: 0.7,
    rfr: 0.3,
    market_price: Some(19.0),
 };
 if let Some(s) = kelly_ratio(&a_option) { println!("{:?}", s); }

Structs

Options

Holds the option data

Functions

bs_price

Calculates the option value with the Black-Scholes formula

expected

Performs a Monte-Carlo analysis with 10000 simulations

kelly_ratio

Calculates the Kelly fraction