Module modus::stock_returns

Portfolio performance

Most calculations of portfolio performance don’t include the whole data or are affected by when an asset is bought and therefore are not suitable for comparison. This module shows performance controlling for those factors.

The function total_returns takes a Portfolio and returns a Result<BTreeMap<String, f64>, StocksError>, StocksError being a custom error enum for the error types that can occur.

Usage:

 let portfolio = Portfolio{portfolio: vec![Equity{ticker: "MSFT".to_string(), buy: Transaction { date: TransactionDate {
        year: 2023,
        month: 2,
        day: 1,
    }, price: 354.0 }, sell: None, quantity: 3 }]};
 if let Ok(s) = total_returns(&portfolio).await { println!("{:?}", s); }

Structs

Portfolio

Holds the historical data about your portfolio

Enums

StocksError

This custom error uses the custom derive macro From to implement the From trait

Functions

total_returns

Returns a Result<BTreeMap<String, f64>, StocksError> where the BTreeMap is composed of a date as key and a percentage gain as value and StocksError is an enum with the different types of Error that might have occurred

Derive Macros

From